+++ backend/src/main/java/mos/common/entity/BaseTimeEntity.java
@@ -1,4 +1,4 @@
-package mos.common;
+package mos.common.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
