+++ backend/src/main/java/mos/participant/entity/Participant.java
@@ -3,7 +3,7 @@ package mos.participant.entity;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import mos.common.BaseTimeEntity;
+import mos.common.entity.BaseTimeEntity;
 import mos.member.entity.Member;
 import mos.mogako.entity.Mogako;
